+++ mne_bids/utils.py
@@ -66,7 +66,8 @@ def _get_ch_type_mapping(fro='mne', to='bids'):
     if fro == 'mne' and to == 'bids':
         mapping = dict(eeg='EEG', misc='MISC', stim='TRIG', emg='EMG',
                        ecog='ECOG', seeg='SEEG', eog='EOG', ecg='ECG',
-                       resp='RESP', bio='MISC', dbs='DBS',
+                       resp='RESP', bio='MISC', dbs='DBS', gsr='GSR',
+                       temperature='TEMP',
                        # NIRS
                        fnirs_cw_amplitude='NIRSCWAMPLITUDE',
                        # MEG channels
@@ -78,7 +79,7 @@ def _get_ch_type_mapping(fro='mne', to='bids'):
     elif fro == 'bids' and to == 'mne':
         mapping = dict(EEG='eeg', MISC='misc', TRIG='stim', EMG='emg',
                        ECOG='ecog', SEEG='seeg', EOG='eog', ECG='ecg',
-                       RESP='resp',
+                       RESP='resp', GSR='gsr', TEMP='temperature',
                        # NIRS
                        NIRSCWAMPLITUDE='fnirs_cw_amplitude',
                        NIRS='fnirs_cw_amplitude',